requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.5